// BankatiProject_Front-end/src/app/client/models/clientRequest.ts
export interface ClientRequest {
  id?: string;
  agentId: string;
  firstName: string; //
  lastName: string;//
  email: string;//
  address: string;//
  cin: string;//
  birthDate: Date;//
  phoneNumber: string;//
  role: string;
  password: string;
  isFirstLogin: boolean;
  commercialRn: string;//
  image: string;//
  patentNumber: string;//
  isPaymentAccountActivated: boolean;
  typeHissab: string;
  currency: string;
  token: string;
}